requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
pkg-resources==0.0.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==12.6.1
six==1.14.0
tornado==6.0.4
